backend/src/main/java/com/example/firstProject/services/UserService.java
package com.example.firstProject.services;

import com.example.firstProject.exception.UserAlreadyExistsException;
import com.example.firstProject.models.User;
import com.example.firstProject.repositories.UserRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.security.core.userdetails.UserDetails;
import org.springframework.security.core.userdetails.UserDetailsService;
import org.springframework.security.core.userdetails.UsernameNotFoundException;
import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
import org.springframework.stereotype.Service;

import java.util.Optional;
import java.util.Set;

// Service class for managing User entities
@Service
public class UserService implements UserDetailsService {

    @Autowired
    UserRepository userRepository;

    // adds a new user to the users table
    public void addUser(User user) throws UserAlreadyExistsException {
        Optional<User> existingUser = userRepository.findByUsername(user.getUsername());
        if (existingUser.isPresent()) {
            throw new UserAlreadyExistsException("User already exists!");
        }
        user.setPassword(new BCryptPasswordEncoder().encode(user.getPassword()));
        user.setRoles(Set.of("USER"));
        userRepository.save(user);
    }

    // loads user-specific data for authentication
    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
        User user = userRepository.findByUsername(username)
                .orElseThrow(() -> new UsernameNotFoundException("User not found: " + username));
        return org.springframework.security.core.userdetails.User
                .withUsername(user.getUsername())
                .password(user.getPassword())
                .authorities(user.getAuthorities())
                .build();
    }
}
